+++ app/src/main/java/com/algorithms/home/Challenge.kt
@@ -1,0 +1,9 @@
+package com.algorithms.home
+
+import androidx.annotation.ColorRes
+
+internal class Challenge(
+    val content: String,
+    @ColorRes val backgroundColor: Int,
+    val launchResource: Int? = null,
+)