+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2024-02-27
 
 ### Added
+- Home section and challenge items.
 - Theme and resources configuration.
 - Documentation files such as CHANGELOG and README.